buildenv/win32-static: add protobuf build script.
This commit also includes a Python script that can convert
centralized .sln dependencies to .vcxproj per-project
dependencies.

This script is needed to build a pristine protobuf straight
from the official zip files, because protobuf ships MSVS 2005
project files. We want -- or really, MSVS 2010 wants -- MSVS 2010
project files.

Most of the conversion to the MSVS 2010 format can be done
using VCUpgrade.exe, found in MSVS 2010 Express, alongside
some search-and-replace work in the .sln file.  Unfortunately,
this method cannot convert dependencies found in the .sln file
to the MSVS 2010 per-project dependencies, because VCUpgrade.exe
works on a per-project basis (it doesn't touch the .sln at all!)

To remedy this, we run our own vs-sln-convert-to-per-project-deps.py
script on the .sln, which then processes the old-style depdency info
from the .sln file and writes new-style per-project dependency
information it to the VCUpgrade.exe-upgraded .vcxproj files.

Ideally, this could have been avoided if we had "devenv.exe /upgrade"
available. However, that's unfortunately only available in MSVS 2010
Pro, which we'd like not to depend on for the build.