buildenv/win32-static: build Qt 5 and modify build as necessary.
This adds Qt 5.1 to the win32-static build environment.

Qt 5 uses zlib in QtCore, and mariadbclient previously bundled a zlib
installation. This meant symbol clashes.

To remedy this, we now build zlib as part of the build environment,
and tell Qt and mariadb to link against our build environment copy.

Since mariadbclient is also used by Qt 4, we also modify the Qt 4
build to use the build environment's zlib library.