buildenv/osx-universal: add -search_paths_first to Qt 4's OpenSSL linker flags.
The -search_paths_first flag forces ld to search each lib directory
for *both* a .dylib and an .a, instead of first searching each directory
for a .dylib, and failing that, searching them for an .a-version instead.

Without this change, our static OpenSSL libs are not considered, and Qt
(and deps) will link against the dynamic system-shipped OpenSSL libraries
instead.