buildenv/win32-static: add support for building with Visual Studio 2013.
This adds support (but does not enable by default) building with
Microsoft Visual Studio 2013.

It includes patches for libsndfile 1.0.25 to build with MSVS2013's
newly gained C99 support. This means we're now able to get debug
symbols for all of our dependencies, and that the Windows build
of Mumble is now fully built with MSVC.

Qt 5 does not yet build with Visual Studio 2013 because we're
building with the v120_xp toolset. It seems that Qt 5 expects
the Windows 8 SDK instead of the Windows 7 SDK that we're using
in the v120_xp XP compatibility mode.