tools/sign-msi.py: use output filename as product description when signing the output .MSI.
This fixes a major source of confusion when installing Mumble's
MSI files.

Prior to this change, the UAC prompt during the Mumble
installation would ask users:

    Do you want to allow the following program to
    install software on your computer?

    Program name:        a4fdda0.msi
    Verified publisher:  <signer's name>
    [...]

This prompt is very confusing, since the MSI file listed
in the UAC dialog *is not* the one the user actually ran.
Instead, it's a temporary name used as the default by
signtool.exe/osslsigncode.

This change modifies sign-msi.py to pass the correct
flags to signtool.exe and osslsigncode to make the
UAC dialog a little more user friendly. Instead of
the random hex-string, we now set the product description
to be the basename of the output MSI file.

This changes the program name in the UAC dialog to be
the name of the output MSI, such as 'mumble-1.2.5.msi'.