tools: in plugin_replacement.py, skip DLLs w/o debug info and keep the build's DLL.
We accidently deployed plugins that have no debug info in them:
https://github.com/mumble-voip/mumble/issues/1214

Add a rule to upload the build's DLLs in that case, to allow overwriting
DLLs without debug information.

We already have a rule where we overwrite DLLs if a PDB can't be
fetched from the symbol store, so this new rule seems like a sensible
addition.