mumble-build: force curl to output fetched file to the basename of the URL.
Just to make sure. I'd expect cURL to do that already here, but since
we use the '-L' flags (which means cURL will follow redirects), we're
not guaranteed to get the filename we expect if we simply use '-O'.

There are probably also other cases that can trigger unwanted
filenames, such as "Content-Disposition: attachment; filename=bla".