mumble-build: add various length checks for the urls array.
In 'vet', check that the '.build' file has a 'urls' array set
before performing its URL check.

In 'fetch', if no urls array is present, print an error and exit,
suggesting the '.build' file to override the fetch function.

In 'verify', don't perform verification if the 'urls' array is empty
or not declared.