tools/fixup-prl-files.py, buildenv/win32-static/qt4.build: change fixup-prl-files.py to work on the install directory; call it after 'nmake install'.
Qt seemingly overwrites the prl files during 'nmake install', which
caused our '.prl' changes to be reverted to the version prior to our
rewriting.

This change addresses that by making the fixup-prl-files.py script
modify the installDir. This is already what the comment in the header
of the script says, so it's all good!

The qt4.build script in buildenv/1.3.x/win32-static is also modified
to call the fixup-prl-files.py script after doing 'nmake install'.