buildenv/1.3.x/win32-static: build Qt 5 without SSE2 for the regular Windows x86 build.
Since Qt 5.3.0, Qt has removed runtime detection for SSE2, and has
begun requiring SSE2 in the default configuration.

Mumble has always only required SSE for Windows, and we don't
want to change that in 1.3.x.

Maybe for 1.4.x. Or maybe we'll tell folks to use x64 instead,
and keep the x86 build around as a legacy option.