buildenv/1.2.x/osx-universal: work around xzcat error in libogg build.
The 'xzcat' utility would, on our 10.6 builder machine, complain
with the following line when trying to 'xzcat' the
libogg-1.3.2.tar.xz file:

    "xzcat: libogg-1.3.2.tar.xz: Is a directory, skipping"

I haven't digged too deep into this, because it's probably limited
to OS X 10.6 or similar old versions of Darwin.

So instead of fixing it at the root, let's just work around it
by not allowing xzcat to interpret the file itself: we'll just
stream it to xzcat via stdin.