gyp, buildenv/1.3.x/win32: remove gyp, use CMake instead for libsndfile - the only user of gyp in the tree.
We can't use gyp anymore, because we use a Python 3 installation.

It's more convenient to use CMake for us, since Gyp is the only
program we use that requires Python 3.

Furthermore, future releases of libsndfile have a CMake-based build
that should also build on MSVC. However, it seems unmaintained at
present. We should probably volunteer to get it into shape.