buildscripts/1.2.x/mumble-win32.cmd: use MumbleNoMergeModule instead of MumbleNoMergeModuleDir.
The installer as always used "MumbleNoMergeNodule".

Without this change, the resulting installer will include both:

 - Individual MSVCRT DLLs
 - Merge modules for the MSVCRT DLLs

Merge modules are problematic, because they often cause "files in use"
warnings from the Windows Installer.

With this change in place, we simply bundle the MSVCRT DLLs,
and forego the merge modules, just as initially intended.