buildenv/1.3.x/win32-static: do not use jom in Qt 5 recipe.
We're seeing very weird errors from Jom on the builders.

Disable it for now.

When inside our buildscript for the build env, in
Jenkins, it fails consistently.

When outside, it doesn't.

We don't know why yet. But we need to update the build environments,
so we'll have to investigate the failures later.