buildenv/1.2.x/win32-static: ensure all of our DLLs are in the PATH.
We got build failures on the Windows builder because uic.exe
in the Qt build was crashing.

It turns out it was attempting to locate zlib1.dll, and it couldn't.

Windows searches for DLLs in PATH, so ensure all the DLLs we use
throughout the build are in there.

We should also consider some automatic means of checking whether
non-system DLLs that we use are actually *from* %MUMBLE_PREFIX%.
It turned out that the build did work on my Windows machine, but
only because it had found a zlib1.dll in c:\Ruby193. Ouch!