Add additional cleanup step to cleanup-buildenv-build-dir.py
Our build environments also create lots of pdbs that might be
useful during the creation of the environment (or inconvenient
to not create) but that are not needed later on. A big example
of these are pdbs for the tests the libraries built.

As we do not want to have to distribute these with our build
environments this patch contains a list of globs that try to
match the biggest known instances of PDBs we do not need later
on.

This patch also adds a number of commandline arguments like
verbose and dry-run to be able to interact with the script
with more confidence during testing and operation.