Properly return error codes for errors in ifs in mumble-win*.cmd
Turns out everything in a if "branch" gets expanded by cmd as soon as the
branch is entered. This means that if errorlevel gets changed in during the
branch because our conditional command failed the following exit call already
had its error code expanded to 0. We can work around this by either using
delayed expansion on the errorlevel in these branches or by checking for errors
after the branches. This patch implements the latter variant. Mostly because we
can (as soon as we have to calls in the branch delayed expansion would be the
only way to go) and because it the impact is smaller at this point. Might switch
over later.