buildenv/1.3.x/win32-static: update libsndfile to 1.0.28.
This updates libsndfile to 1.0.28.

Instead of our own homegrown CMakeLists.txt, we now tweak the one
shipped with libsndfile.

We drop various patches related to CVEs that are not relevant for
1.0.28.

Finally, we drop our big MSVC compatibility patch, which was needed
for MSVC < 2015, where the implementation of vsnprintf and snprintf
in MSVC's CRT were not C99 compatible. Instead of our big patch, we
add a build-time #error if we're trying to build libsndfile on a
verison of MSVC < 2015.