buildenv/1.2.x/win32: add protobuf patch to fix MapDirectory test failure.
We've seen odd behavior from Windows when trying to open files that end in
trailing slashes.

Sometimes, Windows ignores trailing slashes in filenames.

I.e., if thre's a file

   C:\asdf\foo

Windows sometimes allows you to open it as "asdf\\foo\\".
Sometimes it doesn't.

For me, it works on my C: drive, but not my D: drive.

We've previously speculated that this was due to junction points,
but since the root of my D: drive behaves the same, I don't think
that's the culprit.

For now, let's just apply this patch such that protobuf doesn't
expect opening files with a trailing slash to always fail.

Fixes mumble-voip/mumble-releng#65
Updates mumble-voip/mumble-releng#44
Updates mumble-voip/mumble-releng#5