Auto-detect Visual Studio installation (most recent first)
Having to set the VS version manually is not documented in the README.
Having specified an invalid version (not installed VS) fails quite late,
when trying to build boost - there is no check we do when actually
preparing the environment, and calling vscvars.
As a result, the behavior and configuration is quite unexpected.

Instead of specifying the VS version manually, in the prep.cmd file,
the script now checks for the existence of the vscvars script files
from most recent supported VS version first, and uses the first it finds.