Revert "Auto-detect Visual Studio installation (most recent first)" in 1.3.x tree
This reverts commit c7c1e6884e1c04ccd32c1db4852d4252d9d541a8
and merge 97af9d6572ae60f64002e677f715b4eb7a189cc7.

The 1.4.x tree will use the new auto-detection.
The 1.3.x scripts as they are do not work with MSVC 14.1, so it is a bad
idea to auto-detect and use it by default.
Instead, we stay on 14.0 on the 1.3.x configuration.